2 int casts

+++ examples/polybench_hls/adi.py
@@ -88,12 +88,12 @@ def adi(ttype, TSTEPS, N):
                 q[i, 0] = v[0, i]
                 for j in range(1, N - 1):
                     p[i, j] = int(-c / (a * p[i, j - 1] + b))
-                    q[i, j] = int((
+                    q[i, j] = int(int(
                         -d * u[j, i - 1]
-                        + (1.0 + 2.0 * d) * u[j, i]
+                        + (1 + 2 * d) * u[j, i]
                         - f * u[j, i + 1]
                         - a * q[i, j - 1]
-                    ) / (a * p[i, j - 1] + b))
+                    ) / int(a * p[i, j - 1] + b))
 
                 v[N - 1, i] = int(1.0)
                 for j_rev in range(N - 1):
